requirements.txt
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.19
decorator==5.2.1
dnspython==2.8.0
exceptiongroup==1.3.1
executing==2.2.1
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.7.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
mysql-connector-python==9.5.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.2.2
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymongo==4.15.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
